Move quantile tests into data_access.rs
