+++ tests/histogram.rs
@@ -3,7 +3,6 @@
 extern crate hdrsample;
 extern crate num;
 extern crate rand;
-extern crate ieee754;
 
 use self::rand::Rng;
 
@@ -13,7 +12,6 @@ use std::borrow::Borrow;
 use std::cmp;
 use std::fmt;
 use num::Saturating;
-use ieee754::Ieee754;
 
 macro_rules! assert_near {
     ($a: expr, $b: expr, $tolerance: expr) => {{
@@ -920,189 +918,4 @@ fn subtract_underflow_guarded_by_per_value_count_check() {
     h2.record_n(1, 100).unwrap();
 
     assert_eq!(SubtractionError::SubtrahendCountExceedsMinuendCount, h.subtract(h2).unwrap_err());
-}
-
-#[test]
-fn quantile_2_values() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    h.record(1).unwrap();
-    h.record(2).unwrap();
-
-    assert_eq!(1, h.value_at_quantile(0.25));
-    assert_eq!(1, h.value_at_quantile(0.5));
-
-    let almost_half = 0.5000000000000001;
-    let next = 0.5000000000000002;
-    // one ulp apart
-    assert_eq!(almost_half, 0.5_f64.next());
-    assert_eq!(next, almost_half.next());
-
-    assert_eq!(1, h.value_at_quantile(0.5));
-    // ideally this would return 2, not 1
-    assert_eq!(1, h.value_at_quantile(almost_half));
-    assert_eq!(2, h.value_at_quantile(next));
-}
-
-#[test]
-fn quantile_5_values() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    h.record(1).unwrap();
-    h.record(2).unwrap();
-    h.record(2).unwrap();
-    h.record(2).unwrap();
-    h.record(2).unwrap();
-
-    assert_eq!(2, h.value_at_quantile(0.25));
-    assert_eq!(2, h.value_at_quantile(0.3));
-}
-
-
-#[test]
-fn quantile_20k() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    for i in 1..20_001 {
-        h.record(i).unwrap();
-    }
-
-    assert_eq!(20_000, h.count());
-
-    assert!(h.equivalent(19961, h.value_at_quantile(0.99805)));
-}
-
-#[test]
-fn quantile_large_numbers() {
-    let mut h = Histogram::<u64>::new_with_bounds(20_000_000, 100_000_000, 5).unwrap();
-    h.record(100_000_000).unwrap();
-    h.record(20_000_000).unwrap();
-    h.record(30_000_000).unwrap();
-
-    assert!(h.equivalent(20_000_000, h.value_at_quantile(0.5)));
-    assert!(h.equivalent(30_000_000, h.value_at_quantile(0.5)));
-    assert!(h.equivalent(100_000_000, h.value_at_quantile(0.8333)));
-    assert!(h.equivalent(100_000_000, h.value_at_quantile(0.8334)));
-    assert!(h.equivalent(100_000_000, h.value_at_quantile(0.99)));
-}
-
-#[test]
-fn value_at_quantile_matches_pctile_iter_sequence() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    let lengths = vec![1, 5, 10, 50, 100, 500, 1_000, 5_000, 10_000, 50_000, 100_000];
-
-    for length in lengths {
-        h.reset();
-
-        for i in 1..(length + 1) {
-            h.record(i).unwrap();
-        }
-
-        assert_eq!(length, h.count());
-
-        let iter = h.iter_percentiles(1000);
-
-        for iter_val in iter {
-            let calculated_value = h.value_at_quantile(iter_val.quantile());
-            let v = iter_val.value();
-
-            assert_eq!(v, calculated_value,
-                       "len {} iter quantile {} q count {} iter val {} -> {} calc val {} -> {}",
-                       length, iter_val.quantile(), iter_val.quantile() * length as f64, v, h.highest_equivalent(v), calculated_value, h.highest_equivalent(calculated_value));
-        }
-    }
-}
-
-#[test]
-fn value_at_quantile_matches_value_sequence() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    let lengths = vec![1, 5, 10, 50, 100, 500, 1_000, 5_000, 10_000, 50_000, 100_000];
-
-    for length in lengths {
-        h.reset();
-
-        for i in 1..(length + 1) {
-            h.record(i).unwrap();
-        }
-
-        assert_eq!(length, h.count());
-
-        for v in 1..(length + 1) {
-            let quantile = v as f64 / length as f64;
-            let calculated_value = h.value_at_quantile(quantile);
-            if !h.equivalent(v, calculated_value) {
-                assert_eq!(h.highest_equivalent(v), calculated_value,
-                         "len {} quantile {} q count {} actual {} -> {} calc {} -> {}",
-                         length, quantile, quantile * length as f64, v, h.highest_equivalent(v), calculated_value, h.highest_equivalent(calculated_value));
-            }
-        }
-    }
-}
-
-#[test]
-fn value_at_quantile_matches_pctile_iter_random() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-
-    // random u64s tend to be pretty darn big, so percentile calculations have to scan more.
-    let lengths = vec![1, 5, 10, 50, 100, 500, 1_000, 5_000, 10_000];
-
-    let mut rng = rand::weak_rng();
-
-    for length in lengths {
-        h.reset();
-
-        for _ in 1..(length + 1) {
-            h.record(rng.gen()).unwrap();
-        }
-
-        assert_eq!(length, h.count());
-
-        let iter = h.iter_percentiles(1000);
-
-        for iter_val in iter {
-            let calculated_value = h.value_at_quantile(iter_val.quantile());
-            let v = iter_val.value();
-
-            assert_eq!(v, calculated_value,
-                       "len {} iter quantile {} q count {} iter val {} -> {} calc val {} -> {}",
-                       length, iter_val.quantile(), iter_val.quantile() * length as f64, v, h.highest_equivalent(v), calculated_value, h.highest_equivalent(calculated_value));
-        }
-    }
-}
-
-#[test]
-fn value_at_quantile_matches_value_random() {
-    let mut h = Histogram::<u64>::new_with_bounds(1, u64::max_value(), 3).unwrap();
-    let mut values = Vec::new();
-
-    let lengths = vec![1, 5, 10, 50, 100, 500, 1_000, 5_000, 10_000];
-
-    let mut rng = rand::weak_rng();
-
-    for length in lengths {
-        h.reset();
-
-        for _ in 1..(length + 1) {
-            let v = rng.gen();
-            h.record(v).unwrap();
-            values.push(v);
-        }
-
-        values.sort();
-
-        assert_eq!(length, h.count());
-
-        for (index, &v) in values.iter().enumerate() {
-            let quantile = (index + 1) as f64 / length as f64;
-            let calculated_value = h.value_at_quantile(quantile);
-            if !h.equivalent(v, calculated_value) {
-                // TODO this fails quickly
-//                assert_eq!(h.highest_equivalent(v), calculated_value,
-//                           "len {} quantile {} q count {} actual {} -> {} calc {} -> {}",
-//                         length, quantile, quantile * length as f64, v, h.highest_equivalent(v), calculated_value, h.highest_equivalent(calculated_value));
-            }
-        }
-    }
 }
